Merge feat/contributing to dev (#5)
* docs(readme): Adds README template

wearehive has some pretty good suggestions on how to make a README, so I've copied their template
for our use.

* docs(readme): Adds README template

wearehive has some pretty good suggestions on how to make a README, so I've copied their template
for our use. Also merged with dev.

* docs(readme): Added README content

This is the initial draft of the README, which lacks info on deployment, tests, database, and
configuration. Everything else is basically there.

* docs(contributing): Added Code of Conduct

A code of conduct is critical to supporting an open-source ecosystem. I've implemented the
established Contributor Covenant, which I feel is a great basis to start with.

* docs(contributing): Added CONTRIBUTING and templates

It's best to streamline the contribution process. This commit adds a detail explanation of how to
contribute, as well as templates for Issues and Pull Requests.

* docs(Contributing): Format/Organization fix

Moved the CoC to the .github folder, and removed a rogue table.

* docs(contributing): Spelling/phrasing fixes

Makes some adjustments to the phrasing of the generated contributing materials. Also fixes a type in
the README.

* docs(readme): Added README content

This is the initial draft of the README, which lacks info on deployment, tests, database, and
configuration. Everything else is basically there.

* docs(contributing): Added Code of Conduct

A code of conduct is critical to supporting an open-source ecosystem. I've implemented the
established Contributor Covenant, which I feel is a great basis to start with.

* docs(contributing): Added CONTRIBUTING and templates

It's best to streamline the contribution process. This commit adds a detail explanation of how to
contribute, as well as templates for Issues and Pull Requests.

* docs(Contributing): Format/Organization fix

Moved the CoC to the .github folder, and removed a rogue table.

* docs(contributing): Spelling/phrasing fixes

Makes some adjustments to the phrasing of the generated contributing materials. Also fixes a type in
the README.