Merge feat/ci v2 (#7)
* ci(config): Adds test deployment

Merging with dev revealed that we don't have sudo access on CircleCI boxes. This commit adds a test
deployment moving Now to a local install.

* ci(config): Fix spelling error

Will I ever be able to spell the word "checkout" on my first try?

* ci(config): Removed npm install

NPM doesn't play nice with yarn; it was removing dependencies like it always does. Now, we're just
counting on npx to download Now and execute it immediately

* ci(config): Removes cd command

The box couldn't find the bundle.js.map to delete, so this commit foregoes the action of changing
directories.

* ci(config): Fixes another typo

God, I'm an idiot

* ci(config): Reverts config back to normal

Now that the test deployment has occurred successfully, this commit applies the changes to the
staging and production deployment jobs, and removes the test deployment.