build(config): Adds style extraction, scope flattening, directory cleaning
Style extraction in production is important for optimizing in production, and module concatenation
helps increase execution time as well. This commit also adds contenthashes in built filenames, so we
needed to include a webpack plugin for build directory cleaning.  Finally, I moved Webpack Dashboard
back to  just the dev build, because it was giving us mystery errors in production.

#10, #19