ci(circle): Removes sourcemap deletion (#30)
Previously, we were removing sourcemaps during deployment in order to make sure everything was 1MB
or less in file size. Unfortunately, we also added hashes to output names, so we can no longer just
target the map and delete it. However, we no longer need to do this, with the removal of
babel-loader; I've just removed deletion altogether.

#29