chore(eslint): Removes Standard, adds ESLint
The origin of the parsing error was actually from the omission of `plugin:vue/recommended`. Since
Standard can't extend configurations, it's just better for us to remove it and use the base ESLint,
which we were already using for Codacy anyways. This commit removes Standard, Snazzy, and
eslint-plugin-html, adds eslint-config-vue, and updates the config file to use the new plugin.

31